Fix chord symbol for C13 -- see below
We're changing "g:13" to "g:13.11". Why?

Input in \chordmode is just shorthand for pitches; each chord is
converted into raw pitches, e.g. "c:maj7" = "<c e g b>".

This loses authorial intent in the chord naming, and presumably means
the system is brittle in a number of cases where chord names might be
ambiguous.

Besides which, other than maybe education, there's little reason to
input pitches this way, since real music is going to use interesting
voicings instead of just stacking thirds.

Anyway, that means that the ChordNames object has access only to the
pitches, and it names the chords using simple rules.

The problem we have in this song is that \chordmode and ChordNames are
inconsistent in which pitches they consider a dominant 13th quality.

Entering "c:13" in \chordmode omits the 11th degree, under the logic
that the unaltered 11th is not going to be voiced anyway. (True, but
is this for midi or something? I just don't understand why the
software cares.)

But ChordNames distinguishes between a dominant 9th with added 13th
and a dominant 13th, based on whether or not the 11th is included. So
the chord input as "c:13" gets named "C9 13" above the staff.

In theory, we could use `chordNameExceptions` to override the "9 13"
quality name and replace it with "13" as we wish. In practice, it's
easier to just explicitly add the 11th when inputting a dominant 13th,
as in "c:13.11".